reorganize inferences.py file and add klqp
