+++ edward/__init__.py
@@ -2,17 +2,19 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
+from edward import inferences
 from edward import models
 from edward import stats
 from edward import criticisms
-from edward import inferences
 from edward import util
 
 # Direct imports for convenience
+from edward.inferences import Inference, MonteCarlo, VariationalInference, \
+    KLpq, KLqp, MFVI, ReparameterizationKLqp, ReparameterizationKLKLqp, \
+    ReparameterizationEntropyKLqp, ScoreKLqp, ScoreKLKLqp, ScoreEntropyKLqp, \
+    MAP, Laplace
 from edward.models import PyMC3Model, PythonModel, StanModel
 from edward.criticisms import evaluate, ppc
-from edward.inferences import Inference, MonteCarlo, VariationalInference, \
-    MFVI, KLpq, MAP, Laplace
 from edward.util import copy, cumprod, dot, Empty, get_dims, \
     get_session, hessian, kl_multivariate_normal, log_sum_exp, logit, \
     multivariate_rbf, placeholder, rbf, set_seed, tile, to_simplex
