+++ edward/inferences/__init__.py
@@ -1,0 +1,10 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+from edward.inferences.inference import *
+from edward.inferences.klpq import *
+from edward.inferences.klqp import *
+from edward.inferences.map import *
+from edward.inferences.monte_carlo import *
+from edward.inferences.variational_inference import *
