+++ edward/inferences/monte_carlo.py
@@ -1,0 +1,35 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+from edward.inferences.inference import Inference
+
+
+class MonteCarlo(Inference):
+  """Base class for Monte Carlo inference methods.
+  """
+  def __init__(self, latent_vars, data=None, model_wrapper=None):
+    """Initialization.
+
+    Parameters
+    ----------
+    latent_vars : dict of RandomVariable to RandomVariable
+      Collection of random variables to perform inference on. Each
+      random variable is binded to another random variable; the latter
+      will infer the former conditional on data.
+    data : dict, optional
+      Data dictionary which binds observed variables (of type
+      `RandomVariable`) to their realizations (of type `tf.Tensor`).
+      It can also bind placeholders (of type `tf.Tensor`) used in the
+      model to their realizations.
+    model_wrapper : ed.Model, optional
+      A wrapper for the probability model. If specified, the random
+      variables in `latent_vars`' dictionary keys are strings used
+      accordingly by the wrapper. `data` is also changed. For
+      TensorFlow, Python, and Stan models, the key type is a string;
+      for PyMC3, the key type is a Theano shared variable. For
+      TensorFlow, Python, and PyMC3 models, the value type is a NumPy
+      array or TensorFlow tensor; for Stan, the value type is the
+      type according to the Stan program's data block.
+    """
+    super(MonteCarlo, self).__init__(latent_vars, data, model_wrapper)
